Fix libpxbackend not being found
libpxbackend is installed at /usr/lib/x86_64-linux-gnu/libproxy,
and libproxy sets that folder in DT_RUNTIME to be able to find
it. Unfortunately, that only works when the library is installed
in the standard place, not when it is in a snap (like it is the
case with gnome-46-2404), so this change is required.

Fix https://forum.snapcraft.io/t/q-about-migration-to-core24-gnome-46/41100